==> 03_feature_engineering.ipynb <==
import pandas as pd
from datetime import datetime

df = pd.read_csv(r"SBS_Processed_Datasets/02_cleaned_output.csv")

df.info()

df.head()

FEATURE ENGINEERING 

#STEP_1 ADD AGE RELATED FEATURES

# Get the current year
current_year = datetime.now().year

# Calculate car age
df['car_age'] = current_year - df['registered_year']
df['model_age'] = current_year - df['model_year']
df['registration_lag']  = df['registered_year'] - df['model_year'] 
df['normalized_model_age']  = df['car_age'] / (df['model_age']  + 1)


#STEP_2 Add price_per_km

bins = [0, 5000, 10000, 50000, 100000, float('inf')]
labels = ['Unused', 'Very_Low', 'Low', 'Moderate', 'High']  # Remove one label to match bin intervals
df['kms_bins'] = pd.cut(df['kms_driven'], bins=bins, labels=labels, include_lowest=True)


#STEP_3 Create mileage_normalized

# Calculate normalized mileage
df['mileage_normalized'] = df['mileage_kmpl'] / df['mileage_kmpl'].max()

#STEP_4 High Mileage Flag

# high_mileage (binary feature)
high_mileage_threshold = 150000  # Define the threshold for high mileage
df['high_mileage'] = (df['kms_driven'] > high_mileage_threshold).astype(int)

#STEP_5 Multiple Owners

# Flag for cars with multiple previous owners
df['multiple_owners'] = (df['owner_no'] > 1).astype(int)




#STEP_6 Brand Popularity

# Compute average price per brand
brand_avg_price = df.groupby('brand')['price'].mean().to_dict()

# Map the average price to each row
df['brand_popularity'] = df['brand'].map(brand_avg_price)

# Example: Adding brand popularity column
print(df[['brand', 'brand_popularity']].head())


#STEP_7 kms_per_year
df['kms_per_year'] = df['kms_driven'] / (df['car_age'] + 1)


df.head(2)

df.describe()

df.head(2)

df.columns

# Export it
df.to_csv(r"SBS_Processed_Datasets\03_feature_engineered_output.csv",index=False)
